Revert "Creacion de los productos en el dashboard de negocios y alertas de stock bajo"
This reverts commit 27914274d59e4fe81a9ae84a7b1b71de9eb8ac13.

+++ app/src/main/java/com/clay/ecommerce_compose/ui/components/client/business/BusinessHome.kt
@@ -213,7 +213,7 @@ fun UserBusinessComponent(
                                             price = product.price,
                                             imageUrl = product.imageUrl,
                                             quantity = 1,
-                                            stock = product.stock
+                                            stock = product.stock.toInt()
                                         )
                                     )
                                 )
